Start making changes to accept the order message from bakery

+++ src/main/java/be/kdg/sa/warehouse/controller/IngredientController.java
@@ -37,7 +37,7 @@ public class IngredientController {
     }
 
     @PatchMapping("/{id}/quantity")
-    public ResponseEntity<?> changeQuantityOfIngredient(@Valid @PathVariable UUID id, @RequestParam double Quantity){
+    public ResponseEntity<?> changeQuantityOfIngredient(@Valid @PathVariable UUID id, @RequestParam Integer Quantity){
         ingredientService.changeQuantityOfIngredient(id,Quantity);
         return ResponseEntity.status(HttpStatus.CREATED).body("The Quantity was changed to " + Quantity);
     }
